build_pairs: write pairs incrementally and support resuming a cut-off run
stage2's build_pairs.py run got cancelled ~11.6h in (Kaggle session
wall-clock cap, not an error) at 1275/2000 prompts, 951 pairs generated
-- and lost all of it, because the old code only wrote data/dpo_pairs.jsonl
once at the very end. Now writes each pair as a JSON line immediately and
flushes, and tracks progress in a `.progress` sidecar so a resumed run
skips already-processed prompts instead of starting over. Also lower
stage2_dpo.sh's --limit to 500 (measured throughput was ~33s/prompt,
so 2000 prompts was never going to fit in one session) and fix its
skip-if-exists check, which would otherwise treat a partial file as
complete. evaluate.py and build_pairs.py also switch bf16 -> fp16
(T4 has no bf16 tensor cores, same fix already applied to the training
configs).

+++ scripts/evaluate.py
@@ -44,7 +44,8 @@ def main() -> int:
         tok.pad_token = tok.eos_token
     tok.padding_side = "left"
     model = AutoModelForCausalLM.from_pretrained(
-        args.model, torch_dtype=torch.bfloat16, device_map="auto").eval()
+        # T4 (Turing, sm_75) has no bf16 tensor cores; fp16 is native there.
+        args.model, torch_dtype=torch.float16, device_map="auto").eval()
     torch.manual_seed(args.seed)
 
     correct = np.zeros((len(ds), args.k), dtype=bool)
